refactor(v1): move tool/user/server code into a verifiers.v1.mcp subpackage
Split the cramped `tools.py` + `user.py` into `verifiers/v1/mcp/`:
- `server.py` — `ServerBase` (the base authoring class + `run`/`_serve`/`setup`/`setup_task`)
- `toolset.py` — `Toolset` + `ToolsetConfig`
- `user.py` — `User` + `UserConfig`
- `launch.py` — host-side launching: `serve`/`serve_tools`/`serve_shared`/`serve_user`/
  `connect_user` + the runtime mechanics (`serve_in_runtime`, `_install_in_sandbox`, …)
- `__init__.py` — re-exports the public surface

No behavior change. Importers updated (`verifiers.v1`, taskset, rollout, env, interception).
The dependency graph is a clean DAG (server ← toolset/user ← launch). Verified e2e on
subprocess: glossary-v1 1.0, alphabet-sort-v1 user-sim 1.0 (stop=user_completed).
